Update knapsackProb.java
The knapsack function is a dynamic programming function that solves the 0/1 Knapsack problem. The problem is defined as follows: given a set of n items, each with a weight weights[i] and a value values[i], determine the maximum value that can be obtained by selecting a subset of the items with a total weight less than or equal to a given weight limit W. The constraint is that each item can be selected at most once.

The function takes three arguments: the weight limit W, an array of values values, and an array of weights weights. The lengths of the values and weights arrays are assumed to be the same.

The function initializes a 2D array dp of size (n+1) x (W+1), where n is the length of the values array. The dp array is used to store the maximum value that can be obtained with a given number of items and a given weight limit.

The function then iterates over the values and weights arrays using nested loops. For each item i and weight w, the function checks if the weight of the item is less than or equal to the weight limit w. If it is, the function sets dp[i][w] to the maximum of two values:

dp[i-1][w]: This value represents the maximum value that can be obtained without including the current item.
dp[i-1][w-weights[i-1]] + values[i-1]: This value represents the maximum value that can be obtained by including the current item. It is calculated by adding the value of the current item to the maximum value that can be obtained with a weight limit of w - weights[i-1] and a reduced number of items (i-1).
If the weight of the item is greater than the weight limit w, the function sets dp[i][w] to dp[i-1][w], which represents the maximum value that can be obtained without including the current item.

Finally, the function returns the value of dp[n][W], which represents the maximum value that can be obtained with the given weight limit W and all the n items.